Cleanup (improve docs/comments, comment out intentially unused verlog)
We are towards "done" for a handcrafted look at the ALU.

Still need to update python TB to prove specific outcomes are correct
and possible to reach, when the control lines are setup for that mode.

This also documents the specific modes required to encapsulate in
asciidoc.

Nice to have an investigation for layout and control line encoding
specifically for a sky130 cell target.
Nice to look at HA sky130 cell and if providing an additional output
port for AND_AB and XOR_AB from the HA cell, would improve metrics.